requirements.txt
streamlit
pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
docling
google-generativeai